# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NewsLens API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type WinnerMagnitudeSource = (typeof WinnerMagnitudeSource)[keyof typeof WinnerMagnitudeSource];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const WinnerMagnitudeSource = {
  data: "data",
  estimate: "estimate",
} as const;

export interface Winner {
  who: string;
  why: string;
  magnitude: string;
  /**
   * Provenance of the figure in `magnitude`. "data" means it came from the
   * enrichment payload (yfinance / FRED) or from facts extracted out of the
   * article itself. "estimate" means the model inferred it from background
   * knowledge. Clients MUST visually distinguish the two. Absent is treated
   * as "estimate" — the analyze agent demotes anything unrecognised.
   */
  magnitude_source?: WinnerMagnitudeSource;
}

export interface CardAngles {
  /** @nullable */
  india?: string | null;
  /** @nullable */
  usa?: string | null;
  /** @nullable */
  china?: string | null;
}

export interface CardTags {
  domain?: string[];
  geo?: string[];
}

export interface CardSource {
  name: string;
  url: string;
}

export interface MarketQuote {
  label: string;
  /** @nullable */
  symbol?: string | null;
  /** @nullable */
  price?: number | null;
  /** @nullable */
  pct_change_1d?: number | null;
  /** @nullable */
  currency?: string | null;
}

export type CardTimeHorizon = "immediate" | "short" | "long" | null;

export interface Card {
  id: string;
  headline: string;
  summary: string;
  /** @nullable */
  key_number?: string | null;
  winners?: Winner[];
  losers?: Winner[];
  /**
   * Filled only by /cards/personalized. /cards returns null here because cards
   * are generated globally and shared between users.
   * @nullable
   */
  personal_impact?: string | null;
  angles?: CardAngles;
  tags?: CardTags;
  confidence: string;
  /** @nullable */
  time_horizon?: CardTimeHorizon;
  /** Market/macro figures the analysis was anchored to. */
  market_data?: MarketQuote[];
  source?: CardSource;
  /** True for invented demo rows from db/seed.py — clients must label these. */
  is_seed?: boolean;
  created_at: string;
}

export interface CardList {
  cards: Card[];
  total: number;
  page: number;
  limit: number;
  has_more: boolean;
}

export type SwipeInputDirection =
  (typeof SwipeInputDirection)[keyof typeof SwipeInputDirection];

export const SwipeInputDirection = {
  left: "left",
  right: "right",
} as const;

export interface SwipeInput {
  direction: SwipeInputDirection;
  /** @nullable */
  user_id?: string | null;
}

export interface SwipeResult {
  status: string;
}

export interface UserProfile {
  user_id: string;
  /** @nullable */
  income_type?: string | null;
  /** @nullable */
  sector_exposure?: string | null;
  /** @nullable */
  investment_profile?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  companies_of_interest?: string | null;
  /** @nullable */
  created_at?: string | null;
  /** @nullable */
  updated_at?: string | null;
}

export interface UserProfileInput {
  user_id: string;
  /** @nullable */
  income_type?: string | null;
  /** @nullable */
  sector_exposure?: string | null;
  /** @nullable */
  investment_profile?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  companies_of_interest?: string | null;
}

export interface SourceStatus {
  name: string;
  source_type: string;
  /** @nullable */
  last_run?: string | null;
  items_today: number;
  status: string;
}

export interface SourcesStatus {
  sources: SourceStatus[];
  total_cards_live: number;
}

export interface PipelineRunResult {
  status: string;
  /** @nullable */
  message?: string | null;
}

export interface PipelineLog {
  id: string;
  stage: string;
  items_processed: number;
  items_failed: number;
  /** @nullable */
  started_at?: string | null;
  /** @nullable */
  completed_at?: string | null;
  status: string;
  /** @nullable */
  error_log?: string | null;
}

export type ListCardsParams = {
  /**
   * @nullable
   */
  domain?: string | null;
  /**
   * @nullable
   */
  search?: string | null;
  limit?: number;
  page?: number;
};

export type GetUserProfileParams = {
  user_id: string;
};

export type GetPipelineLogsParams = {
  limit?: number;
};