# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NewsLens API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: cards
    description: News card operations
  - name: user
    description: User profile operations
  - name: sources
    description: Source monitoring
  - name: pipeline
    description: Pipeline operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /cards:
    get:
      operationId: listCards
      tags: [cards]
      summary: List news cards
      parameters:
        - name: domain
          in: query
          schema:
            type: ["string", "null"]
        # `search` is declared here but NOT implemented by the backend, which accepts
        # geo + confidence instead (api/routes/cards.py). Left in place because
        # removing it would change the generated client signature; reconcile the two
        # when the spec is next regenerated from FastAPI.
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Card list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardList"

  /cards/personalized:
    get:
      operationId: listPersonalizedCards
      tags: [cards]
      summary: List cards filtered to the user's sectors, with personal impact filled in
      description: >
        Same shape as /cards, with two differences: results are narrowed to the
        sectors declared in the user's profile, and `personal_impact` is populated
        by a read-time personalization pass (cached per card per profile version).
        Returns 404 if the user has no saved profile.
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
        - name: domain
          in: query
          schema:
            type: ["string", "null"]
        - name: geo
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Personalized card list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardList"
        "404":
          description: No profile saved for this user

  /cards/{id}:
    get:
      operationId: getCard
      tags: [cards]
      summary: Get a single card
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Card
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Card"
        "404":
          description: Not found

  /cards/{id}/swipe:
    post:
      operationId: swipeCard
      tags: [cards]
      summary: Swipe a card left or right
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwipeInput"
      responses:
        "200":
          description: Swipe recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SwipeResult"

  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get user profile
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
    put:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"

  /sources/status:
    get:
      operationId: getSourcesStatus
      tags: [sources]
      summary: Get all source statuses
      responses:
        "200":
          description: Sources status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourcesStatus"

  /pipeline/run:
    post:
      operationId: runPipeline
      tags: [pipeline]
      summary: Trigger a pipeline run
      responses:
        "200":
          description: Pipeline triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineRunResult"

  /pipeline/logs:
    get:
      operationId: getPipelineLogs
      tags: [pipeline]
      summary: Get recent pipeline run logs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Pipeline logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PipelineLog"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Winner:
      type: object
      required: [who, why, magnitude]
      properties:
        who:
          type: string
        why:
          type: string
        magnitude:
          type: string
        magnitude_source:
          description: >
            Provenance of the figure in `magnitude`. "data" means it came from the
            enrichment payload (yfinance / FRED) or from facts extracted out of the
            article itself. "estimate" means the model inferred it from background
            knowledge. Clients MUST visually distinguish the two. Absent is treated
            as "estimate" — the analyze agent demotes anything unrecognised.
          type: string
          enum: [data, estimate]

    MarketQuote:
      type: object
      required: [label]
      properties:
        label:
          type: string
        symbol:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        pct_change_1d:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]

    CardAngles:
      type: object
      properties:
        india:
          type: ["string", "null"]
        usa:
          type: ["string", "null"]
        china:
          type: ["string", "null"]

    CardTags:
      type: object
      properties:
        domain:
          type: array
          items:
            type: string
        geo:
          type: array
          items:
            type: string

    CardSource:
      type: object
      required: [name, url]
      properties:
        name:
          type: string
        url:
          type: string

    Card:
      type: object
      required: [id, headline, summary, confidence, created_at]
      properties:
        id:
          type: string
        headline:
          type: string
        summary:
          type: string
        key_number:
          type: ["string", "null"]
        winners:
          type: array
          items:
            $ref: "#/components/schemas/Winner"
        losers:
          type: array
          items:
            $ref: "#/components/schemas/Winner"
        personal_impact:
          description: >
            Filled only by /cards/personalized. /cards returns null here because
            cards are generated globally and shared between users.
          type: ["string", "null"]
        time_horizon:
          type: ["string", "null"]
          enum: [immediate, short, long, null]
        market_data:
          description: >
            Snapshot of the market/macro figures the analysis was anchored to.
            Empty when enrichment found nothing relevant.
          type: array
          items:
            $ref: "#/components/schemas/MarketQuote"
        is_seed:
          description: >
            True for demo rows from db/seed.py. These are invented examples, not
            real news, and clients must label them as such.
          type: boolean
          default: false
        angles:
          $ref: "#/components/schemas/CardAngles"
        tags:
          $ref: "#/components/schemas/CardTags"
        confidence:
          type: string
        source:
          $ref: "#/components/schemas/CardSource"
        created_at:
          type: string

    CardList:
      type: object
      required: [cards, total, page, limit, has_more]
      properties:
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Card"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        has_more:
          type: boolean

    SwipeInput:
      type: object
      required: [direction]
      properties:
        direction:
          type: string
          enum: [left, right]
        user_id:
          type: ["string", "null"]

    SwipeResult:
      type: object
      required: [status]
      properties:
        status:
          type: string

    UserProfile:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: string
        income_type:
          type: ["string", "null"]
        sector_exposure:
          type: ["string", "null"]
        investment_profile:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        companies_of_interest:
          type: ["string", "null"]
        created_at:
          type: ["string", "null"]
        updated_at:
          type: ["string", "null"]

    UserProfileInput:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: string
        income_type:
          type: ["string", "null"]
        sector_exposure:
          type: ["string", "null"]
        investment_profile:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        companies_of_interest:
          type: ["string", "null"]

    SourceStatus:
      type: object
      required: [name, source_type, items_today, status]
      properties:
        name:
          type: string
        source_type:
          type: string
        last_run:
          type: ["string", "null"]
        items_today:
          type: integer
        status:
          type: string

    SourcesStatus:
      type: object
      required: [sources, total_cards_live]
      properties:
        sources:
          type: array
          items:
            $ref: "#/components/schemas/SourceStatus"
        total_cards_live:
          type: integer

    PipelineRunResult:
      type: object
      required: [status]
      properties:
        status:
          type: string
        message:
          type: ["string", "null"]

    PipelineLog:
      type: object
      required: [id, stage, items_processed, items_failed, status]
      properties:
        id:
          type: string
        stage:
          type: string
        items_processed:
          type: integer
        items_failed:
          type: integer
        started_at:
          type: ["string", "null"]
        completed_at:
          type: ["string", "null"]
        status:
          type: string
        error_log:
          type: ["string", "null"]